+++ contrib/python/src/python/pants/contrib/python/checks/tasks/checkstyle/BUILD
@@ -15,6 +15,7 @@ python_library(
     'src/python/pants/base:workunit',
     'src/python/pants/build_graph',
     'src/python/pants/option',
+    'src/python/pants/python',
     'src/python/pants/subsystem',
     'src/python/pants/task',
     'src/python/pants/util:collections',
