+++ contrib/python/src/python/pants/contrib/python/checks/tasks/checkstyle/checkstyle.py
@@ -7,7 +7,6 @@ import sys
 from packaging import version
 from pants.backend.python.interpreter_cache import PythonInterpreterCache
 from pants.backend.python.python_requirement import PythonRequirement
-from pants.backend.python.subsystems.pex_build_util import PexBuilderWrapper
 from pants.backend.python.subsystems.python_setup import PythonSetup
 from pants.backend.python.targets.python_requirement_library import PythonRequirementLibrary
 from pants.backend.python.targets.python_target import PythonTarget
@@ -17,6 +16,7 @@ from pants.base.hash_utils import hash_all
 from pants.base.workunit import WorkUnitLabel
 from pants.build_graph.address import Address
 from pants.option.custom_types import file_option
+from pants.python.pex_build_util import PexBuilderWrapper
 from pants.task.lint_task_mixin import LintTaskMixin
 from pants.task.task import Task
 from pants.util.collections import factory_dict
