+++ contrib/python/src/python/pants/contrib/python/checks/tasks/python_eval.py
@@ -6,8 +6,7 @@ import os
 import pkgutil
 
 from pants.backend.python.interpreter_cache import PythonInterpreterCache
-from pants.backend.python.subsystems.pex_build_util import (PexBuilderWrapper,
-                                                            has_python_requirements,
+from pants.backend.python.subsystems.pex_build_util import (has_python_requirements,
                                                             has_python_sources)
 from pants.backend.python.targets.python_binary import PythonBinary
 from pants.backend.python.targets.python_library import PythonLibrary
@@ -17,6 +16,7 @@ from pants.base.deprecated import deprecated_conditional
 from pants.base.exceptions import TaskError
 from pants.base.generator import Generator, TemplateData
 from pants.base.workunit import WorkUnit, WorkUnitLabel
+from pants.python.pex_build_util import PexBuilderWrapper
 from pants.task.lint_task_mixin import LintTaskMixin
 from pants.util.dirutil import safe_concurrent_creation, safe_mkdir
 from pants.util.memo import memoized_property
