+++ src/python/pants/backend/python/subsystems/BUILD
@@ -20,17 +20,3 @@ python_library(
   ],
   tags = {"partially_type_checked"},
 )
-
-python_tests(
-  name='tests',
-  dependencies=[
-    '3rdparty/python:pex',
-    ':subsystems',
-    'src/python/pants/backend/python/targets',
-    'src/python/pants/build_graph',
-    'src/python/pants/util:contextutil',
-    'src/python/pants/testutil/subsystem',
-    'src/python/pants/testutil:test_base',
-  ],
-  tags = {'partially_type_checked'},
-)
