+++ src/python/pants/backend/python/subsystems/executable_pex_tool.py
@@ -7,7 +7,7 @@ from pex.pex import PEX
 from pex.pex_builder import PEXBuilder
 from pex.pex_info import PexInfo
 
-from pants.backend.python.subsystems.pex_build_util import PexBuilderWrapper
+from pants.python.pex_build_util import PexBuilderWrapper
 from pants.subsystem.subsystem import Subsystem
 from pants.util.dirutil import is_executable, safe_concurrent_creation
