+++ src/python/pants/backend/python/tasks/gather_sources.py
@@ -9,13 +9,13 @@ from pex.pex_builder import PEXBuilder
 from twitter.common.collections import OrderedSet
 
 from pants.backend.python.subsystems.pex_build_util import (
-  PexBuilderWrapper,
   has_python_sources,
   has_resources,
   is_python_target,
 )
 from pants.base.exceptions import TaskError
 from pants.invalidation.cache_manager import VersionedTargetSet
+from pants.python.pex_build_util import PexBuilderWrapper
 from pants.task.task import Task
 from pants.util.dirutil import safe_concurrent_creation
