+++ src/python/pants/python/README.md
@@ -1,4 +1,4 @@
 Python-related functionality needed by the Pants core for bootstrapping.
 
 This is distinct from the Python backend, which provides user-facing
-functionality for building Python code.
+functionality for building Python code (and may reuse this code as appropriate).
